move utils to lib, incl theme
