+++ components/Core.tsx
@@ -1,6 +1,6 @@
 import { Global } from '@emotion/react';
 import { createContext, useState, Fragment, ReactNode } from 'react';
-import { ThemeProvider, useTheme, themeDark, themeLight } from './ui/theme';
+import { ThemeProvider, useTheme, themeDark, themeLight } from '../lib/theme';
 
 const themes = {
   dark: themeDark,
