--- components/ui/README.md
@@ -1,13 +1,0 @@
----
-title: Design System
-path: '/design-system'
-type: design-system
----
-
-# Design System
-
-I created this Design System to showcase my experience and interest in Design Systems.
-
-It is totally open source, but is currently not published for public consumption. Instead, you can explore the code on [GitHub](https://github.com/nathsimpson/nathansimpson.design), or experiment with each of the components in their previews.
-
-You can also see the Figma file [here](https://www.figma.com/file/KD9brps10is6174TXk2YUY/Components?node-id=0%3A1).
