--- components/ui/theme/index.tsx
@@ -1,2 +1,0 @@
-export { ThemeProvider, useTheme, useThemeContext } from './themecontext';
-export * from './themes';
