--- components/ui/utils/index.tsx
@@ -1,5 +1,0 @@
-export { default as hexAlpha } from 'hex-alpha';
-export { forwardRefWithAs } from './forwardRefWithAs';
-export * from './useMediaQuery';
-
-export type ResponsiveProp<T> = T | readonly (T | null)[];
