+++ lib/theme/index.tsx
@@ -1,0 +1,3 @@
+export { ThemeProvider, useThemeContext } from './themecontext';
+export { useTheme } from './useTheme';
+export * from './themes';
