+++ lib/useMediaQuery/README.mdx
@@ -9,14 +9,6 @@ section: hooks
 
 Convenient css-in-js functions for using media queries in layouts.
 
-Reference the breakpoints in theme.
-
-```jsx
-import { useMediaQuery } from '../utils';
-
-const { mq, maxBreak, minBreak } = useMediaQuery();
-```
-
 ## mq
 
 Enables the use of responsive props anywhere.
