+++ lib/useMediaQuery/index.tsx
@@ -1,0 +1,3 @@
+export * from './useMediaQuery';
+
+export type ResponsiveProp<T> = T | readonly (T | null)[];
