+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import { AppProps } from 'next/app';
 import Head from 'next/head';
-import { useTheme } from '../components/ui/theme';
+import { useTheme } from '../lib/theme';
 import { Core } from '../components/Core';
 import { Footer } from '../components/Footer';
