+++ pages/blog/[slug].tsx
@@ -8,8 +8,8 @@ import { Prose } from '../../components/ui/prose';
 import { ContentContainer } from '../../components/Container';
 import { BackButton } from '../../components/BackButton';
 import { Header } from '../../components/Header';
-import type { Post } from '../../lib/posts';
-import { getPostBySlug, getAllPosts } from '../../lib/posts';
+import type { Post } from '../../lib/mdxContent/posts';
+import { getPostBySlug, getAllPosts } from '../../lib/mdxContent/posts';
 import { YouTubeVideo } from '../../components/Blog/YouTubeVideo';
 import { MdxContent } from '../../components/Mdx';
 import { GetStaticProps, InferGetStaticPropsType } from 'next';
